requirements.txt
Django==4.0
python-dotenv==0.19.2
requests==2.22.0
